Lives, Movement, and Heath Counter
Added health counter starting from 3. Respawn point was also added when character loses a life to the center of the screen. Flap was also changed due to if player hits the jump button more than once the character might continue to jump even though they were not when respawning. Kind of like a queue. So changed it to take more responsive jumps.